move schedule workflow modal to modals folder
