+++ packages/frinx-workflow-ui/src/pages/workflow-list/scheduled-workflow/scheduled-workflow-list.tsx
@@ -24,7 +24,7 @@ import FeatherIcon from 'feather-icons-react';
 import PageContainer from '@frinx/workflow-ui/src/common/PageContainer';
 import { usePagination } from '@frinx/workflow-ui/src/common/pagination-hook';
 import callbackUtils from '@frinx/workflow-ui/src/utils/callback-utils';
-import SchedulingModal from './scheduled-workflow-modal/scheduled-workflow-modal';
+import SchedulingModal from '../../../common/modals/scheduled-workflow-modal';
 import { ScheduledWorkflow, StatusType } from '@frinx/workflow-ui/src/helpers/types';
 import useNotifications from '@frinx/workflow-ui/src/hooks/use-notifications';
 import Paginator from '@frinx/workflow-ui/src/common/pagination';
